¡Rename water module to estimators
